Use cloud-archive yoga updates for nova also in xena
We test only for focal, but zed is already on jammy.
And for yoga, the change isn't really a change.

At that point in the script, it will pull in only newer packages related
to qemu, and psycopg2, which isn't used, so the scope of the change is
quite limited.

Here the complete list:
```
librados2
librbd1
libvirt0
openvswitch-common
python3-openvswitch
openvswitch-switch
python3-psycopg2
```