Configure Sentry_SDK
The sentry raven library is long deprecated. The python sentry SDK replaces this library.
To get the same behaviour as with raven
(only enable the logging with selected loggers via the Handlers), one has to initialise
the SDK but manually configure the Logging integration and set the default levels to None.

Avoiding the sentry configuration in each SAPCC upstream based project (e.g. Neutron or Nova) and furthermore
in each driver, this has is part of the loci build process.
During the build a .pth and .py file will be placed in the
site_package path. Within this path, an executable line in a .pth file is run at every Python startup.
In our case, a script configuring sentry if the env variable SENTRY_USE_SDK is
set. As requirements sentry-sdk must be installed.

Co-authored-by: Florian Streibelt  <mail+github@yakshaving.de>