Fix virtualenv installation for Python 3.12
This commit fixes virtualenv installation for Python 3.12 by removing
the bound version 20.7.2. Without this fix, we got the error:
    AttributeError: module 'pkgutil' has no attribute 'ImpImporter'.

At the same time, we keep this binding for older versions because we have
several components that are still running on Python 3.8.

In the future, when we upgrade all our components to newer Python versions,
we have to backport the fix from upstream:
https://github.com/openstack/loci/commit/639122fc1163521a8a37ce3d2a2d8571218c542e